Gaussian2D, change tests to load known signal.
Also, move old test into same file as the new ones.
The TestSymmetricGaussian2DValues is still failing
for some reason.
